This is update PCB_V08 that has brake signal,Data:2018-04-15

+++ nHIOS_PIC16F1828_Sample_pcv08.X/MachineLearing.h
@@ -22,7 +22,10 @@ uchar Machine_N_Learning(void);   //output signal
 
 void Average_First(uchar size,uchar address);
 
-void  convertDecimalToHexa(ulong number);
+void  convertDecimalToHexa(uint number);
+
+
+
 void Average_Second(uchar size_n,uchar address);
